+++ front/src/App.tsx
@@ -16,12 +16,16 @@ import Problem from "./routes/Problem";
 import ErrorPage from "./routes/Error";
 import HomePage from "./routes/Home";
 import Stepper from "./routes/Stepper";
+import Result from "./routes/Result";
+import Profile from "./routes/Profile";
+import LoadingComponent from "./component/Loading/Loading";
 
 function App(): JSX.Element {
     return (
         <RecoilRoot>
             <AnimatePresence>
                 <div className="flex h-screen w-screen flex-col">
+                    <LoadingComponent />
                     <section id="header" className=" h-16 w-full">
                         <NavBar />
                     </section>
@@ -30,6 +34,8 @@ function App(): JSX.Element {
                             <Route path="/" element={<HomePage />} errorElement={<ErrorPage />} />
                             <Route path="/stepper" element={<Stepper />} errorElement={<ErrorPage />} />
                             <Route path="/problem" element={<Problem />} errorElement={<ErrorPage />} />
+                            <Route path="/result" element={<Result />} errorElement={<ErrorPage />} />
+                            <Route path="/profile" element={<Profile />} errorElement={<ErrorPage />} />
                         </Routes>
                     </section>
                 </div>