+++ front/src/api/defaultApi.ts
@@ -2,7 +2,8 @@ import axios from "axios";
 import { DefaultApi } from "../../api/api";
 
 const instance = axios.create();
-const baseURL = "http://15.165.21.53:3000";
+const API_KEY = import.meta.env.API_KEY;
+const baseURL = API_KEY;
 const headers = { "Content-Type": `application/json` };
 
 type postProblemReturn = {